+++ main.py
@@ -1,79 +1,9 @@
 from board import Board
-
-class Piece:
-    def __init__(self, color = 'b'):
-        self.color = color
-        if color.lower() not in 'bw':
-            raise ValueError(f'Color {color} is not supported')
-      
-        
-    
-    def __repr__(self):
-        if self.color.lower() == 'b':
-            return 'X'
-        elif self.color.lower() =='w':
-            return 'O'
-    def same_color(self, piece):
-        try :
-            piece.color
-        except AttributeError('not valid piece'):
-            return False
-
-        return self.color == piece.color
-
-      
-
-class Pawn(Piece):
-    def within_reach(self,x,y):
-        if self.color == 'b':
-            x, y = x*-1, y*-1
-        
-        if x==0 and y == 1:
-            return ('move')
-        elif x == 1 and abs(y)==1 :
-            return ('take')
-        else :
-            return False
-    
+from pieces import Pawn
+from player import Player
 
 
-class Player:
-    """
-    Player collects data on player
-    creates dictionary of posiitons(key) and pieces(values) where pieces of the game are kept and evaluated for game result /valid moves, existence of pieces to be played with
-    args :
-        name:name of the Player
-        color :black or white
-    """
-    def __init__(self, name, color):
-        self.name = name
-        self.pieces = {} 
-        """ enables to work with dictionary methods to retrieve Piece into gaming
-        """
-        if color not in ['b','w']:
-            raise ValueError(f'Color {color} is not valid')
-        else:
-            self.color = color
-            
-            if self.color.lower() == 'b':
-                self.color_repr =  'X'
-            elif self.color.lower() =='w':
-                self.color_repr =  'O'            
-    
-    def __repr__(self):
-        return (f'{self.name} playing {self.color_repr}')
-    
-    def set_piece(self, positions, piece):
-        try:
-            piece.color  = self.color
-        except ValueError:
-            raise ValueError(f'Piece {piece} has no color attribute')
 
-        for position in positions:
-            self.pieces[position]= piece
-    
-    def his_position(self,position):
-        return position in self.pieces.keys()
     
 def vector(board, from_position, to_position):
     try:
